Images-model_comparison/requirements.txt
catboost-1.2.7
cupy-cuda12x==13.3.0
fastrlock==0.8.2
gdown==5.2.0
imblearn==0.0
jupyterlab==3.6.3
lightgbm==4.5.0
matplotlib==3.6.2
numpy==1.26.0
pandas==1.5.0
Pillow==9.4.0
pyyaml==6.0.2
scikit-learn==1.2.2
shap-0.46.0
tqdm==4.67.0
xgboost==2.1.3
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu117